server/functions/auth/requirements.txt
functions-framework==3.*
fastapi==0.104.*
motor==3.*
pymongo==4.*
PyJWT==2.*
python-multipart==0.*
pydantic==2.*
email-validator==2.*
python-dotenv==1.*
a2wsgi==1.*
google-auth==2.*
requests==2.*